DHCT 1.0
Initial public release.

DHCT is a software package to determine and apply corrections to STM
images due to the distortion caused by thermal drift, hysteresis, and
creep. The software was implemented in Matlab 2016a. Check the readme
file for information about image requirements and how to run the
software.

This software is the companion to an article published by our group,
which has been accepted for publication in the Review of Scientific
Instruments as of January 2017.